feature: classifier architecture change & improved valid loss
Changes the architecture of the classifier to a 3 layer network
for outputs and a 2 layer network with the auxilary outputs.
Dropout and ReLU actiation functions were added here as well.
Improved validation loss was also noted at ~85. It appears that
more epochs would improve it even more.

Note that there is higher training loss than valiation loss which
may indicate that the model is overly regularised.

TODO:
	- Train for more iterations.
	- Remove dropout from output classifier and aux output classifier.
	- Retrain all network params not just transfer classifiers