feature: model regularisation reduced. More training epochs
Dropout probability reduced in the outputs classifier and the
auxiliary classifier. This helped to reduce the persistent
difference in loss between the training and validation sets.
The model was also trained for longer - a total of 42 epochs
which further reduced the loss to ~0.82. As validation loss
was continuting to decrease towards the later epochs, it is
suggested to further increase the number of epochs for
training.

TODO:
	- Train with more epochs.
	- Remove dropout entirely from classifiers.
	- Train the weights for the entire network.