feature: removed dropout from classifier and aux classifier
Removed dropout to aleviate the difference between training loss
and validation loss. Model was trained for 100 epochs and a new
best validation loss of ~0.78 was achieved.
There remains a discrepancy between the training loss and the
validation loss with the validation loss generally being lower.
Try removing image transforms as model may still be over-regualised
There appears to some oscillation of the training and validation
loss, so learning rate decay should be applied.

TODO:
	- Learning rate decay
	- Remove data augmentation transforms.
	- Train all parameters of the Inception Net network.
	- Begin task 2 - malignant or benign classification.